# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WAFFER",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://nad-api.squaretech.tech/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/admin/animal-breeds": {
            "get": {
                "tags": [
                    "Admin - Animal Breeds"
                ],
                "description": "Get all animal breeds",
                "operationId": "get_all_animal_breeds",
                "parameters": [
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "animal_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "animal_specie_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Animal Breeds"
                ],
                "description": "Add new animal breed.",
                "operationId": "4111f3489b13ceae79a698edd453a620",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name[ar]",
                                    "category_id",
                                    "animal_type_id",
                                    "animal_specie_id"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "animal_type_id": {
                                        "type": "integer"
                                    },
                                    "animal_specie_id": {
                                        "type": "integer"
                                    },
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/animal-breeds/{id}": {
            "get": {
                "tags": [
                    "Admin - Animal Breeds"
                ],
                "description": "Get animal breed information.",
                "operationId": "show_animal_breed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Animal Breeds"
                ],
                "description": "Edit animal breed.",
                "operationId": "05e02264be608d5bb6658101f695e797",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "animal_type_id": {
                                        "type": "integer"
                                    },
                                    "animal_specie_id": {
                                        "type": "integer"
                                    },
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Animal Breeds"
                ],
                "description": "Delete entered animal breed.",
                "operationId": "delete_animal_breed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}